+++ redwood-layout-composeui/src/test/kotlin/app/cash/redwood/layout/composeui/ComposeFlexContainerTest.kt
@@ -71,7 +71,7 @@ class ComposeFlexContainerTest(
       ),
       listOf(
         LayoutModifier,
-        GrowImpl(1f),
+        GrowImpl(1.0),
         CrossAxisAlignmentImpl(CrossAxisAlignment.Stretch),
       ),
       listOf(
@@ -183,7 +183,7 @@ private inline fun <reified T> cartesianProduct(vararg lists: List<T>): List<Arr
 }
 
 private data class GrowImpl(
-  override val value: Float,
+  override val value: Double,
 ) : Grow
 
 private data class CrossAxisAlignmentImpl(
